Add files via upload
This dataset covers the period for the first quarter of 2019. This project aims to gain insight into city-wide biking trends by analyzing the Divvy trip data. My objectives are to use Python to look into the number of Divvy trips and bikes used, popular start and end stations, the demographic information of Divvy Bikes users, and the days and times Divvy bikes are used.